requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.6
decorator==4.4.2
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.59
llvmlite==0.43.0
MarkupSafe==2.1.5
marshmallow==3.21.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai==1.30.1
openai-whisper==20231117
orjson==3.10.3
packaging==23.2
pillow==10.4.0
proglog==0.1.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
setuptools==70.3.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sympy==1.13.0
tenacity==8.3.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
